+++ lua/john/core/init.lua
@@ -1,0 +1,2 @@
+require "john.core.keymaps"
+require "john.core.options"
