+++ lua/john/core/options.lua
