+++ lua/john/plugins/cmp.lua
