+++ lua/john/plugins/comment.lua
