+++ lua/john/plugins/gitsigns.lua
