+++ lua/john/plugins/hop.lua
