+++ lua/john/plugins/impatient.lua
