+++ lua/john/plugins/init.lua
@@ -1,0 +1,23 @@
+require "john.plugins.lsp" 
+
+require "john.plugins.colorscheme"
+require "john.plugins.cmp"
+require "john.plugins.telescope"
+require "john.plugins.gitsigns"
+require "john.plugins.git-blame"
+require "john.plugins.treesitter"
+require "john.plugins.autopairs"
+require "john.plugins.comment"
+require "john.plugins.nvim-tree"
+require "john.plugins.bufferline"
+require "john.plugins.lualine"
+require "john.plugins.toggleterm"
+require "john.plugins.project"
+require "john.plugins.impatient"
+require "john.plugins.indentline"
+require "john.plugins.alpha"
+require "john.plugins.whichkey"
+require "john.plugins.autocommands"
+require "john.plugins.rust_config"
+require "john.plugins.dap_config"
+require "john.plugins.vim-rest-console"
