+++ lua/john/plugins/lsp/mason.lua
@@ -37,13 +37,13 @@ local opts = {}
 
 for _, server in pairs(servers) do
 	opts = {
-		on_attach = require("user.plugins.lsp.handlers").on_attach,
-		capabilities = require("user.plugins.lsp.handlers").capabilities,
+		on_attach = require("john.plugins.lsp.handlers").on_attach,
+		capabilities = require("john.plugins.lsp.handlers").capabilities,
 	}
 
 	server = vim.split(server, "@")[1]
 
-	local require_ok, conf_opts = pcall(require, "user.plugins.lsp.settings." .. server)
+	local require_ok, conf_opts = pcall(require, "john.plugins.lsp.settings." .. server)
 	if require_ok then
 		opts = vim.tbl_deep_extend("force", conf_opts, opts)
 	end
