+++ lua/john/plugins/lsp/settings/jsonls.lua
