+++ lua/john/plugins/lsp/settings/lua_ls.lua
