+++ lua/john/plugins/lsp/settings/pyright.lua
