+++ lua/john/plugins/monokai.lua
