+++ lua/john/plugins/nvim-tree.lua
