+++ lua/john/plugins/telescope.lua
