--- lua/user/core/init.lua
@@ -1,2 +1,0 @@
-require "user.core.keymaps"
-require "user.core.options"
