--- lua/user/plugins/init.lua
@@ -1,23 +1,0 @@
-require "user.plugins.lsp" 
-
-require "user.plugins.colorscheme"
-require "user.plugins.cmp"
-require "user.plugins.telescope"
-require "user.plugins.gitsigns"
-require "user.plugins.git-blame"
-require "user.plugins.treesitter"
-require "user.plugins.autopairs"
-require "user.plugins.comment"
-require "user.plugins.nvim-tree"
-require "user.plugins.bufferline"
-require "user.plugins.lualine"
-require "user.plugins.toggleterm"
-require "user.plugins.project"
-require "user.plugins.impatient"
-require "user.plugins.indentline"
-require "user.plugins.alpha"
-require "user.plugins.whichkey"
-require "user.plugins.autocommands"
-require "user.plugins.rust_config"
-require "user.plugins.dap_config"
-require "user.plugins.vim-rest-console"
